Add sepolia / mainnet handling and pulled out common code
