+++ .env-example
@@ -1,5 +1,6 @@
 RPC_URL=
 PRIVATE_KEY=
+CHAIN=sepolia
 PATRON_DISTRIBUTOR_ADDRESS=0x7Cea9eB05D34c5F567eC96Dfa085F571A9786811
 PATRON_NFT_ADDRESS=0xB5AF0C7f3885c1007e6699c357566610291585Cb
 PATRON_VESTING_ADDRESS=
